# openapi/openapi.yaml
openapi: 3.0.3
info:
  title: Tunnex API
  version: 0.1.0
  description: |
    Control-plane API for Tunnex.io. This spec is the single source of truth:
    the Go server types/handlers and the TypeScript client are generated from it
    (`make generate`), and drift fails the build (`make generate-check`).
servers:
  - url: /
    description: Same-origin (served behind the edge nginx)

# Endpoints require a session cookie by default; public ones set `security: []`.
security:
  - cookieAuth: []

paths:
  /healthz:
    get:
      operationId: getHealth
      summary: Liveness probe
      description: Reports process liveness. No external dependencies are checked.
      tags: [system]
      security: []
      responses:
        "200":
          description: The service is alive.
          headers:
            X-Request-Id:
              $ref: "#/components/headers/RequestId"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthResponse"
        default:
          $ref: "#/components/responses/Error"

  /api/v1/auth/signup:
    post:
      operationId: signup
      summary: Create a local account (sends a verification email)
      tags: [auth]
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/SignupRequest" }
      responses:
        "202":
          description: Accepted. Returns the same generic result whether or not the email was already registered.
          headers: { X-Request-Id: { $ref: "#/components/headers/RequestId" } }
          content:
            application/json:
              schema: { $ref: "#/components/schemas/GenericMessage" }
        default: { $ref: "#/components/responses/Error" }

  /api/v1/auth/login:
    post:
      operationId: login
      summary: Verify credentials
      description: "S2.1 verifies credentials; S2.2 establishes the session cookie."
      tags: [auth]
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/LoginRequest" }
      responses:
        "200":
          description: Credentials valid.
          headers: { X-Request-Id: { $ref: "#/components/headers/RequestId" } }
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AuthUser" }
        default: { $ref: "#/components/responses/Error" }

  /api/v1/auth/verify-email:
    post:
      operationId: verifyEmail
      summary: Verify an email address with a token
      tags: [auth]
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/TokenRequest" }
      responses:
        "200":
          description: Verified.
          headers: { X-Request-Id: { $ref: "#/components/headers/RequestId" } }
          content:
            application/json:
              schema: { $ref: "#/components/schemas/GenericMessage" }
        default: { $ref: "#/components/responses/Error" }

  /api/v1/auth/password-reset:
    post:
      operationId: requestPasswordReset
      summary: Request a password-reset email
      tags: [auth]
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/PasswordResetRequest" }
      responses:
        "202":
          description: Accepted. Generic result regardless of whether the email exists.
          headers: { X-Request-Id: { $ref: "#/components/headers/RequestId" } }
          content:
            application/json:
              schema: { $ref: "#/components/schemas/GenericMessage" }
        default: { $ref: "#/components/responses/Error" }

  /api/v1/auth/password-reset/confirm:
    post:
      operationId: confirmPasswordReset
      summary: Set a new password using a reset token
      tags: [auth]
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/PasswordResetConfirmRequest" }
      responses:
        "200":
          description: Password updated.
          headers: { X-Request-Id: { $ref: "#/components/headers/RequestId" } }
          content:
            application/json:
              schema: { $ref: "#/components/schemas/GenericMessage" }
        default: { $ref: "#/components/responses/Error" }

  /api/v1/auth/logout:
    post:
      operationId: logout
      summary: Revoke the current session
      description: Idempotent — clears the cookie and revokes the session if present.
      tags: [auth]
      security: []
      responses:
        "204":
          description: Logged out.
          headers: { X-Request-Id: { $ref: "#/components/headers/RequestId" } }
        default: { $ref: "#/components/responses/Error" }

  /api/v1/auth/sso/{provider}/start:
    get:
      operationId: startSsoLogin
      summary: Begin an SSO login (enterprise)
      tags: [sso]
      security: []
      parameters:
        - name: provider
          in: path
          required: true
          schema: { type: string, enum: [google, microsoft] }
        - name: org
          in: query
          required: true
          description: Organization slug whose SSO config to use.
          schema: { type: string }
      responses:
        "200":
          description: The IdP redirect URL for the client to navigate to.
          headers: { X-Request-Id: { $ref: "#/components/headers/RequestId" } }
          content:
            application/json:
              schema: { $ref: "#/components/schemas/SsoRedirect" }
        default: { $ref: "#/components/responses/Error" }

  /api/v1/auth/sso/{provider}/callback:
    get:
      operationId: ssoCallback
      summary: SSO callback (enterprise)
      tags: [sso]
      security: []
      parameters:
        - name: provider
          in: path
          required: true
          schema: { type: string, enum: [google, microsoft] }
        - name: code
          in: query
          required: true
          schema: { type: string }
        - name: state
          in: query
          required: true
          schema: { type: string }
      responses:
        "302":
          description: Authenticated — session cookie set, browser redirected to the app.
          headers:
            Location: { schema: { type: string }, required: true }
        default: { $ref: "#/components/responses/Error" }

  /api/v1/organizations/{orgId}/sso/{provider}:
    parameters:
      - name: orgId
        in: path
        required: true
        schema: { type: string, format: uuid }
      - name: provider
        in: path
        required: true
        schema: { type: string, enum: [google, microsoft] }
    put:
      operationId: setSsoConfig
      summary: Configure an SSO provider for an organization (enterprise)
      tags: [sso]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/SsoConfigRequest" }
      responses:
        "204":
          description: Saved.
          headers: { X-Request-Id: { $ref: "#/components/headers/RequestId" } }
        default: { $ref: "#/components/responses/Error" }

  /api/v1/organizations:
    get:
      operationId: listOrganizations
      summary: List organizations
      description: Lists the organizations the authenticated user belongs to.
      tags: [organizations]
      responses:
        "200":
          description: Organizations.
          headers:
            X-Request-Id:
              $ref: "#/components/headers/RequestId"
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Organization"
        default:
          $ref: "#/components/responses/Error"
    post:
      operationId: createOrganization
      summary: Create an organization
      description: |
        Creates an organization. In the open edition this fails with
        `org_limit_reached` once one organization exists; the enterprise build
        allows multiple.
      tags: [organizations]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateOrganizationRequest"
      responses:
        "201":
          description: The created organization.
          headers:
            X-Request-Id:
              $ref: "#/components/headers/RequestId"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Organization"
        default:
          $ref: "#/components/responses/Error"

  /api/v1/organizations/{orgId}:
    parameters:
      - name: orgId
        in: path
        required: true
        schema:
          type: string
          format: uuid
    get:
      operationId: getOrganization
      summary: Get an organization
      tags: [organizations]
      responses:
        "200":
          description: The organization.
          headers:
            X-Request-Id:
              $ref: "#/components/headers/RequestId"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Organization"
        default:
          $ref: "#/components/responses/Error"
    patch:
      operationId: updateOrganization
      summary: Update organization settings (name only; slug is immutable)
      tags: [organizations]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateOrganizationRequest"
      responses:
        "200":
          description: The updated organization.
          headers:
            X-Request-Id:
              $ref: "#/components/headers/RequestId"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Organization"
        default:
          $ref: "#/components/responses/Error"
    delete:
      operationId: deleteOrganization
      summary: Soft-delete an organization
      tags: [organizations]
      responses:
        "204":
          description: Deleted.
          headers:
            X-Request-Id:
              $ref: "#/components/headers/RequestId"
        default:
          $ref: "#/components/responses/Error"

components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: tunnex_session

  headers:
    RequestId:
      description: Correlation ID for this request; also appears in server logs.
      schema:
        type: string
      required: true

  schemas:
    HealthResponse:
      type: object
      required: [status, service]
      additionalProperties: false
      properties:
        status:
          type: string
          enum: [ok]
          description: Liveness status.
        service:
          type: string
          description: Reporting service name.
          example: tunnex-api
        request_id:
          type: string
          description: Correlation ID for this request.

    Organization:
      type: object
      required: [id, name, slug, created_at, updated_at]
      additionalProperties: false
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        slug:
          type: string
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time

    CreateOrganizationRequest:
      type: object
      required: [name, slug]
      additionalProperties: false
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 200
        slug:
          type: string
          minLength: 1
          maxLength: 63
          pattern: "^[a-z0-9]+(-[a-z0-9]+)*$"

    UpdateOrganizationRequest:
      type: object
      required: [name]
      additionalProperties: false
      description: Only the name is mutable; the slug is fixed at creation.
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 200

    GenericMessage:
      type: object
      required: [message]
      additionalProperties: false
      properties:
        message: { type: string }

    SsoRedirect:
      type: object
      required: [redirect_url]
      additionalProperties: false
      properties:
        redirect_url: { type: string }

    SsoConfigRequest:
      type: object
      required: [client_id, client_secret, enabled]
      additionalProperties: false
      properties:
        client_id: { type: string, minLength: 1, maxLength: 500 }
        client_secret: { type: string, minLength: 1, maxLength: 500 }
        enabled: { type: boolean }

    AuthUser:
      type: object
      required: [id, email, email_verified]
      additionalProperties: false
      properties:
        id: { type: string, format: uuid }
        email: { type: string, format: email }
        email_verified: { type: boolean }

    SignupRequest:
      type: object
      required: [email, password]
      additionalProperties: false
      properties:
        email: { type: string, format: email, maxLength: 320 }
        password: { type: string, minLength: 12, maxLength: 200 }
        name: { type: string, maxLength: 200 }

    LoginRequest:
      type: object
      required: [email, password]
      additionalProperties: false
      properties:
        email: { type: string, format: email, maxLength: 320 }
        password: { type: string, maxLength: 200 }

    TokenRequest:
      type: object
      required: [token]
      additionalProperties: false
      properties:
        token: { type: string, minLength: 1, maxLength: 200 }

    PasswordResetRequest:
      type: object
      required: [email]
      additionalProperties: false
      properties:
        email: { type: string, format: email, maxLength: 320 }

    PasswordResetConfirmRequest:
      type: object
      required: [token, password]
      additionalProperties: false
      properties:
        token: { type: string, minLength: 1, maxLength: 200 }
        password: { type: string, minLength: 12, maxLength: 200 }

    Error:
      type: object
      description: |
        Standard error envelope for every non-2xx response. Chosen over bare
        RFC 7807 so the correlation `request_id` is a first-class field the SPA
        and CLI can surface directly; `code` is stable for programmatic handling.
      required: [error]
      additionalProperties: false
      properties:
        error:
          type: object
          required: [code, message]
          additionalProperties: false
          properties:
            code:
              type: string
              description: Stable, machine-readable error code.
              example: not_found
            message:
              type: string
              description: Human-readable explanation.
            request_id:
              type: string
              description: Correlation ID; match against server logs.
            details:
              type: array
              description: Field-level problems (e.g. validation failures).
              items:
                $ref: "#/components/schemas/ErrorDetail"

    ErrorDetail:
      type: object
      required: [message]
      additionalProperties: false
      properties:
        field:
          type: string
          description: Dotted path to the offending field, when applicable.
        message:
          type: string
          description: What is wrong with this field.

  responses:
    Error:
      description: Error response.
      headers:
        X-Request-Id:
          $ref: "#/components/headers/RequestId"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"